+++ .gitignore
@@ -3,3 +3,13 @@
 /vendor
 /web/packages.json
 /web/p
+
+###> symfony/framework-bundle ###
+/.env.local
+/.env.local.php
+/.env.*.local
+/config/secrets/prod/prod.decrypt.private.php
+/public/bundles/
+/var/
+/vendor/
+###< symfony/framework-bundle ###
