+++ README.md
@@ -56,8 +56,8 @@ The lookup table is provided as many static JSON files. The entry point to these
 
 ### Installing
 
-1. Make sure you have PDO with sqlite support enabled.
-2. Make sure [`data`](data/) is writable. Do NOT create `data/packages.sqlite`, it will be created automatically.
+1. Make sure you have Composer dependencies installed, including extensions.
+2. Make sure [`data`](data/) is writable. Do NOT create `data/packages.sqlite` manually; use the `migrate` command.
 3. Run `composer install`.
 4. Point your Web server to [`web`](web/). A [`.htaccess`](web/.htaccess) is provided for Apache.
 
@@ -65,6 +65,7 @@ The lookup table is provided as many static JSON files. The entry point to these
 
 The first database fetch may easily take 30-60 minutes, be patient.
 
-1. `bin/cmd refresh`: Query the WordPress.org SVN in order to find new and updated packages.
-2. `bin/cmd update`: Update the version information for packages identified in `1`. Uses the WordPress.org API.
-3. `bin/cmd build`: Rebuild all `.json` files in `web/`.
+0. `bin/console migrate`: Ensure the database schema is up to date with the code.
+1. `bin/console refresh`: Query the WordPress.org SVN in order to find new and updated packages.
+2. `bin/console update`: Update the version information for packages identified in `1`. Uses the WordPress.org API.
+3. `bin/console build`: Rebuild all `.json` files in `web/`.
