+++ composer.json
@@ -1,30 +1,48 @@
 {
     "name": "outlandish/wpackagist",
     "description": "Install and manage WordPress plugins with Composer",
-    "require": {
-        "php": ">=5.3.2",
-        "composer/composer": "1.3.*",
-        "silex/silex": "~1.1",
-        "twig/twig": ">=1.8,<2.0-dev",
-        "symfony/console": "*",
-        "symfony/filesystem":"*",
-        "symfony/twig-bridge": "~2.3",
-        "symfony/form": "~2.3",
-        "symfony/security-csrf": "~2.3",
-        "symfony/locale": "~2.3",
-        "symfony/config": "~2.3",
-        "symfony/translation": "~2.3",
-        "pagerfanta/pagerfanta": "dev-master",
-        "franmomu/silex-pagerfanta-provider": "dev-master",
-        "doctrine/dbal": "2.5.*",
-        "knplabs/console-service-provider": "1.0",
-        "rarst/wporg-client": "dev-master",
-        "guzzlehttp/guzzle-services": "1.0.*"
+    "config": {
+        "sort-packages": true
     },
-    "bin-dir": "bin",
+    "minimum-stability": "stable",
+    "require": {
+        "php": ">=7.2",
+        "ext-json": "*",
+        "ext-pdo": "*",
+        "ext-pdo_sqlite": "*",
+        "ext-simplexml": "*",
+        "doctrine/dbal": "^2.10.2",
+        "guzzlehttp/guzzle-services": "^1.1.3",
+        "pagerfanta/pagerfanta": "^2.1.3",
+        "rarst/wporg-client": "~0.5",
+        "symfony/config": "^5.0",
+        "symfony/console": "^5.0",
+        "symfony/dotenv": "^5.0",
+        "symfony/filesystem": "^5.0",
+        "symfony/flex": "^1.6",
+        "symfony/form": "^5.0",
+        "symfony/orm-pack": "^1.0",
+        "symfony/security-csrf": "^5.0",
+        "symfony/twig-pack": "^1.0",
+        "symfony/yaml": "^5.0"
+    },
+    "require-dev": {
+        "composer/composer": "^1.10.6",
+        "roave/security-advisories": "dev-master"
+    },
+    "conflict": {
+        "silex/silex": "*",
+        "symfony/symfony": "*"
+    },
     "autoload": {
         "psr-4": {
             "Outlandish\\Wpackagist\\": "src/"
+        }
+    },
+    "scripts": {
+        "auto-scripts": {
+            "cache:clear": "symfony-cmd",
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
         }
     }
 }
