+++ config/packages/prod/routing.yaml
@@ -1,0 +1,3 @@
+framework:
+    router:
+        strict_requirements: null
