+++ config/packages/routing.yaml
@@ -1,0 +1,3 @@
+framework:
+    router:
+        utf8: true
