+++ config/routes.yaml
@@ -1,0 +1,3 @@
+#index:
+#    path: /
+#    controller: App\Controller\DefaultController::index
