+++ src/Command/RefreshCommand.php
@@ -2,12 +2,11 @@
 
 namespace Outlandish\Wpackagist\Command;
 
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class RefreshCommand extends Command
+class RefreshCommand extends DbAwareCommand
 {
     protected function configure()
     {
@@ -23,7 +22,7 @@ class RefreshCommand extends Command
             );
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $svn = $input->getOption('svn');
 
@@ -32,30 +31,25 @@ class RefreshCommand extends Command
             'theme'  => 'Outlandish\Wpackagist\Package\Theme',
         );
 
-        /**
-         * @var \PDO $db
-         */
-        $db = $this->getApplication()->getSilexApplication()['db'];
-
-        $updateStmt = $db->prepare('UPDATE packages SET last_committed = :date WHERE class_name = :class_name AND name = :name');
-        $insertStmt = $db->prepare('INSERT INTO packages (class_name, name, last_committed) VALUES (:class_name, :name, :date)');
+        $updateStmt = $this->connection->prepare('UPDATE packages SET last_committed = :date WHERE class_name = :class_name AND name = :name');
+        $insertStmt = $this->connection->prepare('INSERT INTO packages (class_name, name, last_committed) VALUES (:class_name, :name, :date)');
 
         foreach ($types as $type => $class_name) {
-            $url = call_user_func(array($class_name, 'getSvnBaseUrl'));
+            $url = call_user_func([$class_name, 'getSvnBaseUrl']);
             $output->writeln("Fetching full plugin list from $url");
 
-            $xmlLines = array();
+            $xmlLines = [];
             exec("$svn ls --xml $url 2>&1", $xmlLines, $returnCode);
-            if ($returnCode) {
-                $output->writeln('<error>Error from svn command</error>');
+            if ($returnCode > 0) {
+                $output->writeln("<error>Error code $returnCode from svn command</error>");
 
-                return 1; // error code
+                return $returnCode; // error code
             }
             $xml = simplexml_load_string(implode("\n", $xmlLines));
 
             $output->writeln("Updating database");
 
-            $db->beginTransaction();
+            $this->connection->beginTransaction();
             $newCount = 0;
             foreach ($xml->list->entry as $entry) {
                 $date = date('Y-m-d H:i:s', strtotime((string) $entry->commit->date));
@@ -67,11 +61,13 @@ class RefreshCommand extends Command
                     $newCount++;
                 }
             }
-            $db->commit();
+            $this->connection->commit();
 
-            $updateCount = $db->query($s = 'SELECT COUNT(*) FROM packages WHERE last_fetched < last_committed AND class_name = '.$db->quote($class_name))->fetchColumn();
+            $updateCount = $this->connection->query($s = 'SELECT COUNT(*) FROM packages WHERE last_fetched < last_committed AND class_name = '.$this->connection->quote($class_name))->fetchColumn();
 
             $output->writeln("Found $newCount new and $updateCount updated {$type}s");
         }
+
+        return 0;
     }
 }
